Revert "Feat/Static build (#2233)"
This reverts commit 6534e6033ed4f3d4dc523890925e274c5c541db1.
