+++ apps/web/app/[locale]/board/page.tsx
@@ -1,10 +1,25 @@
-import { APPLICATION_DEFAULT_LANGUAGE } from '@app/constants';
-import BoardPage from './component';
+'use client';
 
-export async function generateStaticParams() {
-	return [{ locale: APPLICATION_DEFAULT_LANGUAGE }];
+import { withAuthentication } from 'lib/app/authenticator';
+import { BackdropLoader, Meta } from 'lib/components';
+import dynamic from 'next/dynamic';
+
+const Board = dynamic(() => import('lib/features/integrations/boards'), {
+	ssr: false,
+	loading: () => <BackdropLoader show />
+});
+
+function BoardPage() {
+	return (
+		<>
+			<div className="relative">
+				<Meta title="Board" />
+				<Board />
+			</div>
+		</>
+	);
 }
 
-export default function Page() {
-	return <BoardPage />;
-}
+export default withAuthentication(BoardPage, {
+	displayName: 'BoardIntegrationPage'
+});
