+++ apps/web/app/api/image-assets/upload/[folder]/route.ts
@@ -5,15 +5,12 @@ import { INextParams } from '@app/interfaces';
 
 export async function POST(req: Request, { params }: INextParams) {
 	const res = new NextResponse();
-	const folderParam = params.folder;
-
-	if (!folderParam) {
-		return;
-	}
 
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
 	if (!user) return NextResponse.json({}, { status: 401 });
+
+	const folderParam = params.folder as string;
 
 	const form = await req.formData();
