+++ apps/web/app/api/integration-tenant/[id]/route.ts
@@ -1,20 +1,17 @@
-import { INextParams } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import { deleteIntegrationTenantRequest } from '@app/services/server/requests';
 import { NextResponse } from 'next/server';
 
-export async function DELETE(req: Request, { params }: INextParams) {
+export async function DELETE(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-	const integrationId = params.id;
-
-	if (!integrationId) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId, organizationId } = await authenticatedGuard(req, res);
 	if (!user) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
 
-	const response = await deleteIntegrationTenantRequest(integrationId, tenantId, organizationId, access_token);
+	const { id } = params;
 
-	return $res(response.data);
+	if (id) {
+		const response = await deleteIntegrationTenantRequest(id as string, tenantId, organizationId, access_token);
+
+		return $res(response.data);
+	}
 }
