+++ apps/web/app/api/invite/[id]/route.ts
@@ -11,12 +11,12 @@ import { NextResponse } from 'next/server';
 
 export async function GET(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-	if (!params.id) {
-		return NextResponse.json({}, { status: 400 });
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 	if (!user) return NextResponse.json({}, { status: 401 });
+
+	if (params.id) {
+		return NextResponse.json({}, { status: 400 });
+	}
 
 	const { data } = await getMyInvitationsRequest(tenantId, access_token);
 
@@ -25,14 +25,14 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
 
 export async function DELETE(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-	const invitationId = params.id;
-
-	if (!invitationId) {
-		return NextResponse.json({}, { status: 400 });
-	}
-
 	const { $res, user, access_token, tenantId, organizationId, teamId } = await authenticatedGuard(req, res);
 	if (!user) return NextResponse.json({}, { status: 401 });
+
+	const invitationId = params.id;
+
+	if (params.id) {
+		return NextResponse.json({}, { status: 400 });
+	}
 
 	await removeTeamInvitationsRequest({
 		bearer_token: access_token,
@@ -63,7 +63,7 @@ export async function PUT(req: Request, { params }: { params: { id: string } }) 
 	const { searchParams } = new URL(req.url);
 	const { action } = searchParams as unknown as { action: string };
 
-	if (!invitationId) {
+	if (params.id) {
 		return NextResponse.json({}, { status: 400 });
 	}
