+++ apps/web/app/api/issue-types/[id]/route.ts
@@ -1,22 +1,18 @@
-import { INextParams } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import { deleteIssueTypesRequest, editIssueTypesRequest } from '@app/services/server/requests';
 import { NextResponse } from 'next/server';
 
-export async function PUT(req: Request, { params }: INextParams) {
+export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
 	if (!user) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
 
+	const { id } = params;
 	const body = await req.json();
 
 	const response = await editIssueTypesRequest({
-		id: params.id,
+		id,
 		datas: body,
 		bearer_token: access_token,
 		tenantId
@@ -25,18 +21,16 @@ export async function PUT(req: Request, { params }: INextParams) {
 	return $res(response.data);
 }
 
-export async function DELETE(req: Request, { params }: INextParams) {
+export async function DELETE(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
 	if (!user) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
 
+	const { id } = params;
+
 	const response = await deleteIssueTypesRequest({
-		id: params.id,
+		id,
 		bearer_token: access_token,
 		tenantId
 	});
