+++ apps/web/app/api/organization-projects/[id]/route.ts
@@ -5,18 +5,16 @@ import { NextResponse } from 'next/server';
 
 export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 	if (!user) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+
+	const { id } = params;
 
 	const body = await req.json();
 
 	const response = await editOrganizationProjectsRequest({
 		bearer_token: access_token,
-		id: params.id,
+		id,
 		datas: body,
 		tenantId
 	});
