+++ apps/web/app/api/organization-projects/setting/[id]/route.ts
@@ -1,24 +1,19 @@
-import { INextParams } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 
 import { editOrganizationProjectsSettingsRequest } from '@app/services/server/requests';
 import { NextResponse } from 'next/server';
 
-export async function PUT(req: Request, { params }: INextParams) {
+export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 	if (!user) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
 
+	const { id } = params;
 	const body = await req.json();
 
 	const response = await editOrganizationProjectsSettingsRequest({
-		id: params.id,
 		bearer_token: access_token,
+		id,
 		datas: body,
 		tenantId
 	});
