+++ apps/web/app/api/organization-team-employee/[id]/active-task/route.ts
@@ -1,27 +1,24 @@
-import { INextParams, IOrganizationTeamEmployeeUpdate } from '@app/interfaces';
+import { IOrganizationTeamEmployeeUpdate } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import { updateOrganizationTeamEmployeeActiveTaskRequest } from '@app/services/server/requests';
 import { NextResponse } from 'next/server';
 
-export async function PUT(req: Request, { params }: INextParams) {
+export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
-
 	if (!user) return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
 
+	const { id } = params;
 	const body = (await req.json()) as IOrganizationTeamEmployeeUpdate;
 
-	const response = await updateOrganizationTeamEmployeeActiveTaskRequest({
-		id: params.id,
-		bearer_token: access_token,
-		tenantId,
-		body
-	});
+	if (id) {
+		const response = await updateOrganizationTeamEmployeeActiveTaskRequest({
+			id: id as string,
+			bearer_token: access_token,
+			tenantId,
+			body
+		});
 
-	return $res(response.data);
+		return $res(response.data);
+	}
 }
