+++ apps/web/app/api/public/team/[profileLink]/[teamId]/route.ts
@@ -1,32 +1,27 @@
-import { INextParams } from '@app/interfaces';
 import {
 	getPublicOrganizationTeamMiscDataRequest,
 	getPublicOrganizationTeamRequest
 } from '@app/services/server/requests/public-organization-team';
 import { NextResponse } from 'next/server';
 
-export async function GET(req: Request, { params }: INextParams) {
+export async function GET(req: Request, { params }: { params: { profileLink: string; teamId: string } }) {
 	const { searchParams } = new URL(req.url);
 
-	if (!params.profileLink || !params.teamId) {
-		return;
-	}
-
-	const type = searchParams.get('type') as string;
+	const { profileLink, teamId } = params;
+	const { type } = searchParams as unknown as { type: string };
 
 	if (type === 'misc') {
 		const response = await getPublicOrganizationTeamMiscDataRequest({
-			profileLink: params.profileLink,
-			teamId: params.teamId
+			profileLink: profileLink,
+			teamId
 		});
 
 		return NextResponse.json(response.data);
 	}
 
 	const response = await getPublicOrganizationTeamRequest({
-		profileLink: params.profileLink,
-		teamId: params.teamId
+		profileLink: profileLink,
+		teamId
 	});
-
 	return NextResponse.json(response.data);
 }
