+++ apps/web/app/api/tags/[id]/route.ts
@@ -1,23 +1,18 @@
-import { INextParams, ITaskLabelsCreate } from '@app/interfaces';
+import { ITaskLabelsCreate } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import { deleteTaskLabelsRequest, editTaskLabelsRequest } from '@app/services/server/requests';
 import { NextResponse } from 'next/server';
 
-export async function DELETE(req: Request, { params }: INextParams) {
+export async function DELETE(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return NextResponse.json({}, { status: 400 });
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
-	if (!user) {
-		return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
-	}
+	if (!user) return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
+
+	const { id } = params;
 
 	const response = await deleteTaskLabelsRequest({
-		id: params.id,
+		id,
 		bearer_token: access_token,
 		tenantId
 	});
@@ -25,23 +20,17 @@ export async function DELETE(req: Request, { params }: INextParams) {
 	return $res(response.data);
 }
 
-export async function PUT(req: Request, { params }: INextParams) {
-	const res = new NextResponse();
-
-	if (!params.id) {
-		return NextResponse.json({}, { status: 400 });
-	}
-
+export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const body = (await req.json()) as unknown as ITaskLabelsCreate;
-
+	const res = new NextResponse();
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
-	if (!user) {
-		return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
-	}
+	if (!user) return NextResponse.json({ error: 'unauthorized' }, { status: 401 });
+
+	const { id } = params;
 
 	const response = await editTaskLabelsRequest({
-		id: params.id,
+		id,
 		datas: body,
 		bearer_token: access_token,
 		tenantId
