+++ apps/web/app/api/task-priorities/[id]/route.ts
@@ -1,44 +1,38 @@
-import { INextParams, ITaskPrioritiesCreate } from '@app/interfaces';
+import { ITaskPrioritiesCreate } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import { deleteTaskPrioritiesRequest, editTaskPrioritiesRequest } from '@app/services/server/requests';
 import { NextResponse } from 'next/server';
 
-export async function PUT(req: Request, { params }: INextParams) {
+export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return NextResponse.json({}, { status: 400 });
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
 	if (!user) return $res('Unauthorized');
 
+	const { id } = params;
+
 	const datas = (await req.json()) as unknown as ITaskPrioritiesCreate;
 
 	const response = await editTaskPrioritiesRequest({
-		id: params.id,
-		bearer_token: access_token,
+		id,
 		datas,
+		bearer_token: access_token,
 		tenantId
 	});
 
 	return $res(response.data);
 }
 
-export async function DELETE(req: Request, { params }: INextParams) {
+export async function DELETE(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return NextResponse.json({}, { status: 400 });
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
 	if (!user) return $res('Unauthorized');
 
+	const { id } = params;
+
 	const response = await deleteTaskPrioritiesRequest({
-		id: params.id,
+		id,
 		bearer_token: access_token,
 		tenantId
 	});
