+++ apps/web/app/api/task-related-issue-types/[id]/route.ts
@@ -1,4 +1,4 @@
-import { INextParams, ITaskRelatedIssueTypeCreate } from '@app/interfaces';
+import { ITaskRelatedIssueTypeCreate } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import {
 	deleteTaskRelatedIssueTypeRequest,
@@ -6,46 +6,36 @@ import {
 } from '@app/services/server/requests/task-related-issue-type';
 import { NextResponse } from 'next/server';
 
-export async function PUT(req: Request, { params }: INextParams) {
+export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
-	if (!user) {
-		return $res('Unauthorized');
-	}
+	if (!user) return $res('Unauthorized');
+
+	const { id } = params;
 
 	const datas = (await req.json()) as unknown as ITaskRelatedIssueTypeCreate;
 
 	const response = await editTaskRelatedIssueTypeRequest({
-		id: params.id,
-		bearer_token: access_token,
+		id,
 		datas,
+		bearer_token: access_token,
 		tenantId
 	});
 
 	return $res(response.data);
 }
 
-export async function DELETE(req: Request, { params }: INextParams) {
+export async function DELETE(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
-	if (!user) {
-		return $res('Unauthorized');
-	}
+	if (!user) return $res('Unauthorized');
+
+	const { id } = params;
 
 	const response = await deleteTaskRelatedIssueTypeRequest({
-		id: params.id,
+		id,
 		bearer_token: access_token,
 		tenantId
 	});
