+++ apps/web/app/api/task-versions/[id]/route.ts
@@ -1,46 +1,38 @@
-import { INextParams, ITaskVersionCreate } from '@app/interfaces';
+import { ITaskVersionCreate } from '@app/interfaces';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import { deleteTaskVersionRequest, editTaskVersionRequest } from '@app/services/server/requests/task-version';
 import { NextResponse } from 'next/server';
 
-export async function PUT(req: Request, { params }: INextParams) {
+export async function PUT(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
 	if (!user) return $res('Unauthorized');
 
-	const datas = (await req.json()) as ITaskVersionCreate;
+	const { id } = params;
+
+	const datas = (await req.json()) as unknown as ITaskVersionCreate;
 
 	const response = await editTaskVersionRequest({
-		id: params.id,
-		bearer_token: access_token,
+		id,
 		datas,
+		bearer_token: access_token,
 		tenantId
 	});
 
 	return $res(response.data);
 }
 
-export async function DELETE(req: Request, { params }: INextParams) {
+export async function DELETE(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
-
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
 
-	if (!user) {
-		return $res('Unauthorized');
-	}
+	if (!user) return $res('Unauthorized');
+
+	const { id } = params;
 
 	const response = await deleteTaskVersionRequest({
-		id: params.id,
+		id,
 		bearer_token: access_token,
 		tenantId
 	});
