+++ apps/web/app/api/user/[id]/route.ts
@@ -1,43 +1,30 @@
-import { INextParams } from '@app/interfaces';
 import { IUser } from '@app/interfaces/IUserData';
 import { authenticatedGuard } from '@app/services/server/guards/authenticated-guard-app';
 import { getTaskCreator, updateUserAvatarRequest } from '@app/services/server/requests';
 import { deleteUserRequest } from '@app/services/server/requests/user';
 import { NextResponse } from 'next/server';
 
-export async function GET(req: Request, { params }: INextParams) {
+export async function GET(req: Request, { params }: { params: { id: string } }) {
 	const res = new NextResponse();
 
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token } = await authenticatedGuard(req, res);
+	if (!user) return $res('Unauthorized');
 
-	if (!user) {
-		return $res('Unauthorized');
-	}
+	const { id: userId } = params;
 
 	const { data } = await getTaskCreator({
-		userId: params.id,
+		userId: userId as string,
 		bearer_token: access_token
 	});
 
 	return $res(data);
 }
 
-export async function PUT(req: Request, { params }: INextParams) {
+export async function PUT(req: Request) {
 	const res = new NextResponse();
 
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
-
-	if (!user) {
-		return $res('Unauthorized');
-	}
+	if (!user) return $res('Unauthorized');
 
 	const body = (await req.json()) as unknown as IUser;
 
@@ -53,18 +40,11 @@ export async function PUT(req: Request, { params }: INextParams) {
 	return $res(response.data);
 }
 
-export async function DELETE(req: Request, { params }: INextParams) {
+export async function DELETE(req: Request) {
 	const res = new NextResponse();
 
-	if (!params.id) {
-		return;
-	}
-
 	const { $res, user, access_token, tenantId } = await authenticatedGuard(req, res);
-
-	if (!user) {
-		return $res('Unauthorized');
-	}
+	if (!user) return $res('Unauthorized');
 
 	const response = await deleteUserRequest({
 		id: user.id,
