+++ apps/web/app/hooks/features/useTimeSlot.ts
@@ -31,8 +31,9 @@ export function useTimeSlots(hasFilter?: boolean) {
 				todayEnd,
 				todayStart
 			}).then((response) => {
-				if (response.data && Array.isArray(response.data)) {
-					setTimeSlots(response.data[0]?.timeSlots || []);
+				if (response.data) {
+					// @ts-expect-error
+					setTimeSlots(response.data[0].timeSlots);
 				}
 			});
 		} else setTimeSlots([]);
