+++ apps/web/middleware.ts
@@ -1,5 +1,4 @@
 import {
-	APPLICATION_DEFAULT_LANGUAGE,
 	DEFAULT_APP_PATH,
 	DEFAULT_MAIN_PATH,
 	PROTECTED_APP_URL_PATHS,
@@ -31,7 +30,7 @@ export const config = {
 
 export async function middleware(request: NextRequest) {
 	const nextIntlMiddleware = createMiddleware({
-		defaultLocale: APPLICATION_DEFAULT_LANGUAGE,
+		defaultLocale: 'en',
 		locales: ['en', 'de', 'ar', 'bg', 'zh', 'nl', 'de', 'he', 'it', 'pl', 'pt', 'ru', 'es', 'fr'],
 		// pathnames,
 		localePrefix: 'as-needed'
