+++ apps/web/next.config.js
@@ -8,8 +8,6 @@ const isProduction = process.env.NODE_ENV === 'production';
 
 const isSentryEnabled = isProduction && process.env.SENTRY_DSN;
 
-const BUILD_OUTPUT_MODE = process.env.NEXT_BUILD_OUTPUT_TYPE;
-
 const sentryConfig = isSentryEnabled && {
 	sentry: {
 		// For all available options, see: https://docs.sentry.io/platforms/javascript/guides/nextjs/manual-setup/
@@ -33,7 +31,7 @@ const sentryConfig = isSentryEnabled && {
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-	output: ['standalone', 'export'].includes(BUILD_OUTPUT_MODE) ? BUILD_OUTPUT_MODE : undefined,
+	output: process.env.NEXT_BUILD_OUTPUT_TYPE === 'standalone' ? 'standalone' : undefined,
 	reactStrictMode: false,
 	swcMinify: true,
 	webpack: (config, { isServer }) => {
